% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_encoding_c.R
\name{prepare.model_residual_trajectory_code}
\alias{prepare.model_residual_trajectory_code}
\title{Residual trajectory encode categorical variables.}
\usage{
\method{prepare}{model_residual_trajectory_code}(treatmentplan, dframe,
  ...)
}
\arguments{
\item{treatmentplan}{a model_residual_trajectory_code coder}

\item{dframe}{data frame to be encoded.}

\item{...}{not used, force arguments to be bound by name}
}
\value{
data frame encoding the categorical columns specified in treatmentplan
}
\description{
Residual encode.
}
