% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classificaton_residuals.R
\name{calculate_residual_classification_trajectory}
\alias{calculate_residual_classification_trajectory}
\title{Build residual classification trajectory.}
\usage{
calculate_residual_classification_trajectory(data, fit_predict, vars, ...,
  dep_var, dep_target = TRUE,
  cross_plan = vtreat::kWayStratifiedY(nrow(data), 3, data,
  data[[dep_var]] == dep_target), cl = NULL)
}
\arguments{
\item{data}{The data.frame of data to fit.}

\item{fit_predict}{A function with signature fit_predict(train_data, vars, dep_var, dep_target, application_data) that returns a matrix with one row of predictions per row of appication_data, and an ordered set of columns of predictions.}

\item{vars}{character vector, explanatory variable names.}

\item{...}{not used, force arguments to be bound by name}

\item{dep_var}{character, the name of dependent variable.}

\item{dep_target}{scalar, the value considered to be the target category of dep_var.}

\item{cross_plan}{a vtreat-style cross validation plan for data rows (list of disjoint tran/app lists where app partitions the data rows).}

\item{cl}{parallel cluster for processing}
}
\value{
a matrix with the same number of rows as data representing the cross-validated modeling residual trajectories.
}
\description{
Build a cross-validated residual trajectory for a model.
}
\examples{

data <- iris
vars <- c("Sepal.Length", "Petal.Length",
          "Sepal.Width", "Petal.Width")
dep_var <- "Species"
dep_target <- "versicolor"
resids <- calculate_residual_classification_trajectory(
  data = data,
  vars = vars,
  fit_predict = xgboost_fit_predict_c,
  dep_var = dep_var,
  dep_target = dep_target
)

}
\keyword{internal}
